Widen syscall return values to 64 bits

+++ sys/swi.c
@@ -7,7 +7,7 @@
 
 #include <stdint.h>
 
-typedef uint32_t (*syscall_fn_t)(uint32_t r0, uint32_t r1, uint32_t r2, uint32_t r3);
+typedef uint64_t (*syscall_fn_t)(uint32_t r0, uint32_t r1, uint32_t r2, uint32_t r3);
 
 #define SYSCALLS_TABLE 1
 
@@ -18,7 +18,7 @@ static const syscall_fn_t general[4096] = {
 static const syscall_fn_t priv[4096] = {
 };
 
-uint32_t swi_handler(uint32_t r0, uint32_t r1, uint32_t r2, uint32_t r3)
+uint64_t swi_handler(uint32_t r0, uint32_t r1, uint32_t r2, uint32_t r3)
 {
   register uint32_t r7 __asm__ ("r7");
   uint32_t num = r7;
@@ -29,7 +29,7 @@ uint32_t swi_handler(uint32_t r0, uint32_t r1, uint32_t r2, uint32_t r3)
   } else {
     fn = general[num & 4095];
   }
-  uint32_t ret = 0;
+  uint64_t ret = 0;
   if (fn != NULL) ret = (*fn)(r0, r1, r2, r3);
   return ret;
 }
